Add files via upload
Play a number guessing game. A random integer from 0 to 9 is generated and you have 3 guesses to figure out the actual number. If you guess wrong the first time, you will get a hint on whether the actual number is lower or greater than your guess. If you guess wrong the second time, you will get a hint on whether the actual number is lower or greater than your guess and if your guess is less than 3 away from the actual number or if your guess is more than 3 away from the actual number.